Render Path Re-Write I ditched the original render path, in its entirety. The filtering chain I was building is going to be junked in favor of something using PostFX (a shader-based filter library).
In the process of making that change, though, I re-wrote the render path so that it can render the output at *any* resolution. Long term, this is a "Big Deal", in that it means that the screen size is currently no longer the limiting factor for output resolution. Currently, the `saveImage` method saves the image at 4 times the displayed resolution.